lib: refactor operation-flows with task
Big code refactoring: pass task object deep into each operation floe,
so it may be used as a context object for future asynchronous actions,
such as flush-buffers. This major code change is the first part of code
re-writes which is required in order to enable multiple threads to
execute concurrently. In particular, do not force reader threads to wait
for completion of flush operation from previous writes (which is still
the case with current code base).

Signed-off-by: Shachar Sharon <synarete@gmail.com>