fs: keep vnodes live via refcnt
Setting a vnode dirty is not enough to keep it alive in cache during
operational cycle: in case of memory pressure it may be flushed and
later on become evicted from cache if not pinned with ref-count
explicitly. This, in turn, may yield all sort of voodoo bugs as the
actual memory instance is still alive on the call stack.

Signed-off-by: Shachar Sharon <synarete@gmail.com>