fs: enable file-type2
Allow having blocks-only tree-mapping files (a.k.a file-type2). The
actual file-type is determined by thew control-flags of its parent
directory inode. For regular-file, the type is set upon inode's creation
and can not be modified during the file's lifetime. For directories, the
file-type flag is set on its inode and used to resolve the type of its
child regular files.

Signed-off-by: Shachar Sharon <synarete@gmail.com>